+++ tests/contrib/asyncio/test_helpers.py
@@ -7,15 +7,13 @@ from ddtrace.contrib.asyncio import helpers
 from .utils import AsyncioTestCase, mark_asyncio
 
 
-@pytest.mark.skipif(
-    CONTEXTVARS_IS_AVAILABLE,
-    reason='only applicable to legacy asyncio integration'
-)
+@pytest.mark.skipif(CONTEXTVARS_IS_AVAILABLE, reason="only applicable to legacy asyncio integration")
 class TestAsyncioHelpers(AsyncioTestCase):
     """
     Ensure that helpers set the ``Context`` properly when creating
     new ``Task`` or threads.
     """
+
     @mark_asyncio
     def test_set_call_context(self):
         # a different Context is set for the current logical execution
@@ -32,24 +30,24 @@ class TestAsyncioHelpers(AsyncioTestCase):
             # the ctx is available in this task
             ctx = self.tracer.get_call_context()
             assert 1 == len(ctx._trace)
-            assert 'coroutine' == ctx._trace[0].name
+            assert "coroutine" == ctx._trace[0].name
             return ctx._trace[0].name
 
-        self.tracer.trace('coroutine')
+        self.tracer.trace("coroutine")
         # schedule future work and wait for a result
         delayed_task = helpers.ensure_future(future_work(), tracer=self.tracer)
         result = yield from asyncio.wait_for(delayed_task, timeout=1)
-        assert 'coroutine' == result
+        assert "coroutine" == result
 
     @mark_asyncio
     def test_run_in_executor_proxy(self):
         # the wrapper should pass arguments and results properly
         def future_work(number, name):
             assert 42 == number
-            assert 'john' == name
+            assert "john" == name
             return True
 
-        future = helpers.run_in_executor(self.loop, None, future_work, 42, 'john', tracer=self.tracer)
+        future = helpers.run_in_executor(self.loop, None, future_work, 42, "john", tracer=self.tracer)
         result = yield from future
         assert result
 
@@ -62,10 +60,10 @@ class TestAsyncioHelpers(AsyncioTestCase):
             # span is here to keep the parenting
             ctx = self.tracer.get_call_context()
             assert 0 == len(ctx._trace)
-            assert 'coroutine' == ctx._current_span.name
+            assert "coroutine" == ctx._current_span.name
             return True
 
-        span = self.tracer.trace('coroutine')
+        span = self.tracer.trace("coroutine")
         future = helpers.run_in_executor(self.loop, None, future_work, tracer=self.tracer)
         # we close the Context
         span.finish()
@@ -80,10 +78,10 @@ class TestAsyncioHelpers(AsyncioTestCase):
             # the ctx is available in this task
             ctx = self.tracer.get_call_context()
             assert 0 == len(ctx._trace)
-            child_span = self.tracer.trace('child_task')
+            child_span = self.tracer.trace("child_task")
             return child_span
 
-        root_span = self.tracer.trace('main_task')
+        root_span = self.tracer.trace("main_task")
         # schedule future work and wait for a result
         task = helpers.create_task(future_work())
         result = yield from task
