+++ tests/contrib/asyncio/test_tracer_safety.py
@@ -10,6 +10,7 @@ class TestAsyncioSafety(AsyncioTestCase):
     bad traces are produced but the ``Context`` object will not
     leak memory.
     """
+
     def setUp(self):
         # Asyncio TestCase with the wrong context provider
         super(TestAsyncioSafety, self).setUp()
@@ -22,27 +23,27 @@ class TestAsyncioSafety(AsyncioTestCase):
         assert ctx is not None
         # test that it behaves the wrong way
         task = asyncio.Task.current_task()
-        task_ctx = getattr(task, '__datadog_context', None)
+        task_ctx = getattr(task, "__datadog_context", None)
         assert task_ctx is None
 
     @mark_asyncio
     def test_trace_coroutine(self):
         # it should use the task context when invoked in a coroutine
-        with self.tracer.trace('coroutine') as span:
-            span.resource = 'base'
+        with self.tracer.trace("coroutine") as span:
+            span.resource = "base"
 
         traces = self.tracer.writer.pop_traces()
         assert 1 == len(traces)
         assert 1 == len(traces[0])
-        assert 'coroutine' == traces[0][0].name
-        assert 'base' == traces[0][0].resource
+        assert "coroutine" == traces[0][0].name
+        assert "base" == traces[0][0].resource
 
     @mark_asyncio
     def test_trace_multiple_calls(self):
         @asyncio.coroutine
         def coro():
             # another traced coroutine
-            with self.tracer.trace('coroutine'):
+            with self.tracer.trace("coroutine"):
                 yield from asyncio.sleep(0.01)
 
         ctx = self.tracer.get_call_context()
